+++ code/src/components/FaqPage2/FaqText2.css
